# back/central/shared/docs/auth/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio REST para la gestión de reservas multi-restaurante.",
        "title": "Restaurant Reservation API",
        "termsOfService": "https://ejemplo.com/terminos",
        "contact": {
            "name": "Equipo de Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3050",
    "basePath": "/api/v1",
    "paths": {
...
        "/business-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene una lista de todos los tipos de negocio del sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "business-types"
                ],
                "summary": "Obtener lista de tipos de negocio",
                "responses": {
                    "201": {
                        "description": "Tipos de negocio obtenidos exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo tipo de negocio en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "business-types"
                ],
                "summary": "Crear un nuevo tipo de negocio",
                "parameters": [
                    {
                        "description": "Datos del tipo de negocio a crear",
                        "name": "businessType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BusinessTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tipo de negocio creado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/businesses/configured-resources/{resource_id}/activate": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Activa un recurso configurado para un business específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Resources"
                ],
                "summary": "Activar recurso de business",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID del recurso",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID del business (solo super admin)",
                        "name": "business_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso activado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Sin permisos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Business o recurso no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/businesses/configured-resources/{resource_id}/deactivate": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Desactiva un recurso configurado para un business específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Resources"
                ],
                "summary": "Desactivar recurso de business",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID del recurso",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID del business (solo super admin)",
                        "name": "business_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso desactivado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Sin permisos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Business o recurso no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.BusinessTypeRequest": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de business JWT con el prefijo **Bearer** (para todos los endpoints)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BusinessTokenAuth": {
            "description": "Token principal JWT con el prefijo **Bearer** (solo para /auth/business-token)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}